Multiple things
Made landing page, have two modes pages setup with random word page working, got a logo we need to implement.
